Fix pylint issues.
Skip-unit-tests: true
Skip-fault-injection-test: true
Skip-func-hw-test-medium-verbs-provider: true
Skip-func-hw-test-large: true
Skip-func-hw-test-medium-ucx-provider: false
Allow-unstable-test: true
Test-tag: vm test_dmg_network_scan_basic NetworkFailureTest + ReplayTests

Required-githooks: true
